Moved keydown event listener off the window (#518)
* Moved keydown event listener off the window and onto the image
gallery container. Prevents slides from changing while focused
outside of the slider and using the left and right arrow keys.

Related to #517

* Added useWindowKeyDown prop

When false, keydown events will be listened for on the gallery
rather than the window. Having the listener on the window may
cause the gallery to change at unexpected times (ex: when
moving between characters in an input field).

* Removed package-lock.json

* Removed extra removeEventListener event

* Reverted package-lock.json removal

* Reverted package-lock.json to original version