Image & thumbnail: add loading param
Add “loading” param to all image tags.

“loading” param can be used in image tags to use native lazy load by adding loading=“lazy”. This means browsers will only start loading image when it appears on the screen. Nowadays Google Lighthouse / Insights scores pages negatively if images below the fold are not lazyloaded.

- Add “loading” parameter to item. Will either default to “eager” (=html default, normal behavior), or user can set it to “lazy”
- Add “thumbnailLoading” parameter to item. Same behavior, user can set thumbnails to lazyload
- Add mentiones of these new parameters to readme

I would’ve like to add a test case for this, but the current test command & setup do not seem to work. Something about Jest not liking one of lodash dependencies. I checked this with the example app to confirm the lazyload parameters passed through.